// src - prod/lib/types.ts
export const InvoiceStatus = { ...PrismaInvoiceStatus, REJEITADA: 'REJEITADA' } as const;
export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];
// ...
export interface NoteHistoryEvent extends PrismaNoteHistoryEvent {
  author?: Partial<PrismaUser> | null;
}

export interface FiscalNote extends Omit<PrismaFiscalNote, 'status' | 'amount'> {
  status: InvoiceStatus; // Usa o nosso enum customizado
  amount: Prisma.Decimal | null; // CORREÇÃO: Usa o tipo Decimal do Prisma
  history?: NoteHistoryEvent[];
  attestedBy?: string | null;
  attestedAt?: Date | null;
  observation?: string | null;
  user?: Partial<PrismaUser>; // Include user for accessing email
  creator?: Partial<PrismaUser>; // Adicionado para refletir o schema
  projectTitle?: string | null;
  reportFileName?: string | null;
  reportFileUrl?: string | null;
  deleted: boolean; 
  deletedAt: Date | null; 
}